Use an interface OnScreenWidgetMixin to add extra padding to FootPreference, IntroPreference, and TopIntroPreference
The padding of FooterPreference, IntroPreference, and TopIntroPreference are not enough, so we need to add extra padding for them. We can achieve this by adding interface OnScreenWidgetMixin and letting SettingsPreferenceGroupAdapter to add extra padding to the widgets that implement this interface.

Bug: 409555871
Test: Manually Test the widget paddings are correct
Flag: com.android.settingslib.widget.theme.flags.is_expressive_design_enabled
Change-Id: I41ef7a585779098f10f0a83a6e7c0f6a916fac96
Signed-off-by: HeroBuxx <me@herobuxx.me>